sql/lab-2-partitioning/exercise/requirements.txt
lorem-text==3.0
psycopg2-binary==2.9.9
